integrate last fix for TLPImport.cpp from trunk revision 4730: declare TLPGraphBuilder.version data member as double to avoid further confusion when comparing it
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_5_maint@4732 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/TLPImport.cpp
@@ -86,7 +86,7 @@ namespace tlp {
         std::map<int,Graph *> clusterIndex;
         DataSet *dataSet;
         bool inTLP;
-        float version;
+        double version;
 
         TLPGraphBuilder(Graph *graph, DataSet *dataSet): _graph(graph),
         dataSet(dataSet) {
@@ -101,8 +101,8 @@ namespace tlp {
         bool addString(const std::string& str) {
             // only used to handle the version of tlp file format
             if (!version) {
-                version = (float) atof(str.c_str());
-                return version <= ((float)TLP_VERSION);
+                version = atof(str.c_str());
+                return version <= TLP_VERSION;
             }
             return false;
         }
